Canvas height and width defined
Canvas defined. requestAnimationFrame is used instead of interval because its more accurate.
ship is being moved through an update function which clears the canvas every keystroke.
The moveShip function itself takes the event keycode and moves it left or right, but an operator is used to make sure it doesn't go out the canvas.